several GQL models splitted into files
