+++ gql_events/GraphTypeDefinitions/PresenceTypeGQLModel.py
@@ -1,0 +1,31 @@
+import strawberry as strawberryA
+import datetime
+from typing import Union, List, Annotated, Optional
+from .utils import withInfo, getLoaders
+from ..GraphResolvers import resolveGroupsForEvent, resolvePresencesForEvent
+
+
+@strawberryA.federation.type(keys=["id"], description="""Represents a type of presence (like Present)""")
+class PresenceTypeGQLModel:
+
+    @classmethod
+    async def resolve_reference(cls, info: strawberryA.types.Info, id: strawberryA.ID):
+        loader = getLoaders(info).presencetypes
+        if id is None:
+            return None
+        result = await loader.load(id)
+        if result is not None:
+            result._type_definition = cls._type_definition  # little hack :)
+        return result
+
+    @strawberryA.field(description="""Primary key""")
+    def id(self) -> strawberryA.ID:
+        return self.id
+
+    @strawberryA.field(description="""Name of type (cze)""")
+    def name(self) -> str:
+        return self.name
+
+    @strawberryA.field(description="""Name of type (en)""")
+    def name_en(self) -> str:
+        return self.name_en
