+++ gql_events/GraphTypeDefinitions/__init__.py
@@ -1,4 +1,7 @@
 from .utils import withInfo, getLoaders
 from .EventGQLModel import EventGQLModel
 from .PresenceGQLModel import PresenceGQLModel
+from .EventTypeGQLModel import EventTypeGQLModel
+from .PresenceTypeGQLModel import PresenceTypeGQLModel
+from .InvitationTypeGQLModel import InvitationTypeGQLModel
 from .externals import UserGQLModel, GroupGQLModel
